backend/requirements.txt
# Vita - Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Validation & Serialization
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Utilities
python-dateutil>=2.8.2

# Development
httpx>=0.26.0
